pull out a registrProperty class
