+++ core/core.vcxproj
@@ -646,6 +646,7 @@
     <ClInclude Include="utility\import.h" />
     <ClInclude Include="utility\output.h" />
     <ClInclude Include="utility\registry.h" />
+    <ClInclude Include="propertyBag\registryProperty.h" />
     <ClInclude Include="utility\strings.h" />
   </ItemGroup>
   <ItemGroup>
@@ -742,6 +743,7 @@
     <ClCompile Include="utility\import.cpp" />
     <ClCompile Include="utility\output.cpp" />
     <ClCompile Include="utility\registry.cpp" />
+    <ClCompile Include="propertyBag\registryProperty.cpp" />
     <ClCompile Include="utility\strings.cpp" />
   </ItemGroup>
   <ItemGroup>
