+++ core/core.vcxproj.filters
@@ -333,6 +333,9 @@
     <ClInclude Include="propertyBag\registryPropertyBag.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="propertyBag\registryProperty.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="utility\strings.cpp">
@@ -582,6 +585,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="propertyBag\registryPropertyBag.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="propertyBag\registryProperty.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
